Struct murmurhash64::MurmurHasher

pub struct MurmurHasher {
    // some fields omitted
}

MurmurHash2 can also be used as the hash algorithm in a HashMap (or similar). It implements the necessary traits.

The new hasher traits are only available since Rust 1.7.0 To use them, enable the hasher feature in your build.

Basic Example

let mut hashmap : HashMap<_, _, MurmurState> = Default::default();
hashmap.insert("abc", 123);
hashmap.insert("def", 456);
assert_eq!(Some(&123), hashmap.get("abc"));
assert_eq!(Some(&456), hashmap.get("def"));

let mut hashmap : HashMap<_, _, RandomMurmurState> = Default::default();
hashmap.insert("abc", 123);
hashmap.insert("def", 456);
assert_eq!(Some(&123), hashmap.get("abc"));
assert_eq!(Some(&456), hashmap.get("def"));

Methods

impl MurmurHasher

fn new() -> MurmurHasher

fn with_seed(seed: u64) -> MurmurHasher